Nearest-Point App
Updated a lot of the math in the files, though they are now surpassed in functionality by the app- made an app with buttons to represent our decision boundaries so that we can experiment with different methods / algorithms. Updating double_curve_model and figure_testing though we don't really use them anymore. Auto mode not implemented yet.